add StandardTable class, its test codes, and the other codes with regard to it

+++ onthego.database/src/main/java/onthego/database/core/tablespace/meta/TablespaceHeader.java
@@ -34,8 +34,8 @@ public interface TablespaceHeader {
 	
 	public void setTableMetaInfo(TableMetaInfo tableMetaInfo);
 
-	long getRecordCount();
+	int getRecordCount();
 
-	void setRecordCount(long recordCount);
+	void setRecordCount(int recordCount);
 
 }
